requirements.txt
anthropic>=0.25.0
requests>=2.31.0
feedparser>=6.0.10
beautifulsoup4>=4.12.0
jinja2>=3.1.0
resend>=0.7.0
python-dotenv>=1.0.0
PyGithub>=2.1.0
schedule>=1.2.0
lxml>=5.0.0
